Re-organize extensions, mixins and functions for album art, strings and persistence
- also fetch album art with compute
