+++ lib/player.dart
@@ -1,4 +1,4 @@
 export 'src/player/player_model.dart';
 export 'src/player/player_service.dart';
-export 'src/player/player_utils.dart';
+export 'src/player/player_mixin.dart';
 export 'src/player/view/player_view.dart';
