+++ lib/src/common/safe_network_image.dart
@@ -113,19 +113,3 @@ class XdgCacheManager extends CacheManager with ImageCacheManager {
 
   XdgCacheManager._() : super(Config(key, fileSystem: _XdgFileSystem(key)));
 }
-
-class UrlStore {
-  static final UrlStore _instance = UrlStore._internal();
-  factory UrlStore() => _instance;
-  UrlStore._internal();
-
-  final _value = <String, String?>{};
-
-  String? put({required String icyTitle, String? imageUrl}) {
-    return _value.containsKey(icyTitle)
-        ? _value.update(icyTitle, (value) => imageUrl)
-        : _value.putIfAbsent(icyTitle, () => imageUrl);
-  }
-
-  String? get(String? icyTitle) => icyTitle == null ? null : _value[icyTitle];
-}
