+++ lib/src/local_audio/view/artist_page.dart
@@ -11,7 +11,6 @@ import '../../../get.dart';
 import '../../../local_audio.dart';
 import '../../../player.dart';
 import '../../../settings.dart';
-import '../../../utils.dart';
 import '../../common/explore_online_popup.dart';
 import '../../l10n/l10n.dart';
 import 'genre_page.dart';
@@ -43,7 +42,7 @@ class ArtistPage extends StatelessWidget with WatchItMixin {
     void onAlbumTap(text) {
       final audios = model.findAlbum(Audio(album: text));
       if (audios?.firstOrNull == null) return;
-      final id = generateAlbumId(audios!.first);
+      final id = audios!.first.albumId;
       if (id == null) return;
 
       Navigator.of(context).push(
